refactor: share ruler apply invocation between hooks
Extract the offline `pnpm exec ruler apply` call into scripts/ruler-lib.mjs
(applyRuler), used by both the postinstall and post-merge hooks, so the
offline invariant and flags live in one place. Each hook keeps its own
trigger logic (hash gate vs git-diff range).
